pulling delete, download and get file API to main

+++ PSI_Project/Controllers/ConspectusController.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.FileProviders;
 
 namespace PSI_Project.Controllers;
 
@@ -7,6 +8,32 @@ namespace PSI_Project.Controllers;
 public class ConspectusController : ControllerBase
 {
     private ConspectusHandler _conspectusHandler = new ConspectusHandler();
+    
+    [HttpGet("get/{conspectusId}")]
+    public IActionResult GetConspectus(string conspectusId)
+    {
+        try
+        {
+            Conspectus? conspectus = _conspectusHandler.GetConspectusById(conspectusId);
+            
+            if (conspectus == null)
+                throw new Exception("file not found");
+            
+            string dirPath = Path.GetDirectoryName(conspectus.Path);
+            string filename = Path.GetFileName(conspectus.Path);
+            
+            IFileProvider provider = new PhysicalFileProvider(dirPath);
+            IFileInfo fileInfo = provider.GetFileInfo(filename);
+            var readStream = fileInfo.CreateReadStream();
+
+            return File(readStream, "application/pdf", filename);
+        }
+        catch (Exception ex)
+        {
+            Console.WriteLine(ex.Message);
+            return NotFound();
+        }
+    }
     
     [HttpGet("list")]
     public IActionResult ListUploadedFiles()
@@ -26,39 +53,47 @@ public class ConspectusController : ControllerBase
             FileStream fileStream = new FileStream(filePath, FileMode.Create);
             formFile.CopyTo(fileStream);
             
-            // saving file to db (conspectus.txt file)
-            Conspectus conspectus = new Conspectus(filePath);
-            _conspectusHandler.UploadConspectus(conspectus);
+            // saving file to conspectus.txt (temp db)
+            _conspectusHandler.UploadConspectus(new Conspectus(filePath));
         }
         
         return Ok(_conspectusHandler.ConspectusList);
     }
 
-    [HttpGet("download/{filename}")]
-    public IActionResult DownloadFile(string filename)
+    [HttpGet("download/{conspectusId}")]
+    public IActionResult DownloadFile(string conspectusId)
     {
-        string filePath = Path.Combine(Directory.GetCurrentDirectory(), "Files", filename);
-
+        // checking if there is conspectus with such id
+        Conspectus? conspectus = _conspectusHandler.GetConspectusById(conspectusId);
+        if (conspectus == null)
+            return NotFound();
+        
+        // checking if the file with such path exists
+        string filePath = conspectus.Path;
         if (System.IO.File.Exists(filePath))
-        {
-            return PhysicalFile(filePath, "application/octet-stream");
-        }
-
+            return PhysicalFile(filePath, "application/pdf");
+        
+        // if there is no file, return not found error
         return NotFound();
     }
 
-    [HttpDelete("delete/{filename}")]
-    public void DeleteFile(string filename)
+    [HttpDelete("delete/{conspectusId}")]
+    public void DeleteFile(string conspectusId)
     {
-        string filePath = Path.Combine(Directory.GetCurrentDirectory(), "Files", filename);
-
+        Conspectus? conspectus = _conspectusHandler.GetConspectusById(conspectusId);
+        if (conspectus == null)
+            return;
+                
+        string filePath = conspectus.Path;
         try
         {
             if (System.IO.File.Exists(filePath))
             {
+                // deleting conspectus from filesystem
                 System.IO.File.Delete(filePath);
-                Conspectus conspectus = new Conspectus(filePath);
-                _conspectusHandler.RemoveConspectus(conspectus);
+                
+                // deleting conspectus from conspectus.txt (temp db)
+                _conspectusHandler.RemoveConspectus(conspectusId);
             }
         }
         catch (IOException ex)
